+++ docs/pages/release_notes.md
@@ -13,6 +13,28 @@ This is a {{ site.pmd.release_type }} release.
 {% tocmaker is_release_notes_processor %}
 
 ### 🚀 New and noteworthy
+#### New: Java 23 Support
+
+This release of PMD brings support for Java 23. There are no new standard language features,
+but a couple of preview language features:
+
+* [JEP 455: Primitive Types in Patterns, instanceof, and switch (Preview)](https://openjdk.org/jeps/455)
+* [JEP 476: Module Import Declarations (Preview)](https://openjdk.org/jeps/476)
+* [JEP 477: Implicitly Declared Classes and Instance Main Methods (Third Preview)](https://openjdk.org/jeps/477)
+* [JEP 482: Flexible Constructor Bodies (Second Preview)](https://openjdk.org/jeps/482)
+
+Note that String Templates (introduced as preview in Java 21 and 22) are not supported anymore in Java 23,
+see [JDK-8329949](https://bugs.openjdk.org/browse/JDK-8329949) for details.
+
+In order to analyze a project with PMD that uses these preview language features,
+you'll need to enable it via the environment variable `PMD_JAVA_OPTS` and select the new language
+version `23-preview`:
+
+    export PMD_JAVA_OPTS=--enable-preview
+    pmd check --use-version java-23-preview ...
+
+Note: Support for Java 21 preview language features have been removed. The version "21-preview"
+are no longer available.
 
 ### 🌟 New and changed rules
 #### New Rules
@@ -23,6 +45,7 @@ This is a {{ site.pmd.release_type }} release.
 * apex-performance
   * [#5139](https://github.com/pmd/pmd/issues/5139): \[apex] OperationWithHighCostInLoop: false negative for triggers
 * java
+  * [#5062](https://github.com/pmd/pmd/issues/5062): \[java] Support Java 23
   * [#5167](https://github.com/pmd/pmd/issues/5167): \[java] java.lang.IllegalArgumentException: \<?\> cannot be a wildcard bound
 * java-bestpractices
   * [#3602](https://github.com/pmd/pmd/issues/3602): \[java] GuardLogStatement: False positive when compile-time constant is created from external constants
@@ -42,6 +65,7 @@ This is a {{ site.pmd.release_type }} release.
   * [#5132](https://github.com/pmd/pmd/issues/5132): \[plsql] TomKytesDespair: XPathException for more complex exception handler
 
 ### 🚨 API Changes
+#### Deprecations
 * pmd-jsp
   * {%jdoc jsp::lang.jsp.ast.JspParserImpl %} is deprecated now. It should have been package-private
     because this is an implementation class that should not be used directly.
@@ -55,6 +79,13 @@ This is a {{ site.pmd.release_type }} release.
 * pmd-visualforce
   * {%jdoc visualforce::lang.visualforce.ast.VfParserImpl %} is deprecated now. It should have been package-private
     because this is an implementation class that should not be used directly.
+
+#### Experimental
+* pmd-java
+  * Renamed `isUnnamedClass()` to {%jdoc !!java::lang.java.ast.ASTCompilationUnit#isSimpleCompilationUnit() %}
+  * {%jdoc java::lang.java.ast.ASTImplicitClassDeclaration %}
+  * {%jdoc !!java::lang.java.ast.ASTImportDeclaration#isModuleImport() %}
+  * {%jdoc !ac!java::lang.java.ast.JavaVisitorBase#visit(java::lang.java.ast.ASTImplicitClassDeclaration,P) %}
 
 ### ✨ External Contributions
 * [#5125](https://github.com/pmd/pmd/pull/5125): \[plsql] Improve merge statement (order of merge insert/update flexible, allow prefixes in column names) - [Arjen Duursma](https://github.com/duursma) (@duursma)